add .evn template
