+++ .env-template
@@ -1,0 +1,12 @@
+// API service port
+PORT = ""
+
+// Database identifier info.
+DB_USER = ""
+DB_PASSWORD = ""
+DB_HOST = ""
+DB_PORT = ""
+DB_DATABASE_NAME = ""
+
+// Ethereum RPC endpoint
+RPC_ENDPOINT = ""
